Removed seperate package-composition test runnabled (moved back into unit-tests).
It was not possible to move the proptest generators into the main library and activate them as a feature for the test build. Even if the feature was enabled for dev-only the cargo publish mechanism would not allow a reference to an unpublished version (even if the unpublished version was the same as the currently present one).

Therefor sadly the prop test generators had to be moved back into the unit tests and the composition test also merged back into the unit tests.